test(ui): split maintenance feature specs
